+++ frontend/public/sw.js
@@ -908,3 +908,4 @@ self.addEventListener("fetch", (event) => {
         })(),
     );
 });
+// Soundspan app build: "V6tKT0ZmMMNn7pr1fCsDc"